+++ CoolProp/REFPROP.h
@@ -45,7 +45,7 @@
 		double rhosatV(double T);
 
 	};
-	bool set_REFPROP_fluid(std::string Ref, double *x);
+	bool set_REFPROP_fluid(std::string Ref, std::vector<double> &x);
 	std::string get_REFPROP_fluid_path();
 
 	double REFPROP(char Output,        char Name1,        double Prop1, char Name2,        double Prop2, char * Ref);