+++ CoolProp/gitrevision.h
@@ -1,1 +1,1 @@
-std::string gitrevision = "fb2c56efd5591f7911c24795cd2d5bf08d6024c5";
+std::string gitrevision = "7299dc06720bd7b5a81f8099438083c465e1fe85";